Solution to Calculate and Print the Area of a Circle

+++ Calculate and Print the Area of a Circle@@6788d5012db25efec946f943/code.c
@@ -2,7 +2,7 @@
 int main(){
     int radius;
     scanf("%d",&radius);
-    int ar = 3.14*radius*radius;
+    float ar = 3.14*radius*radius;
     printf("Area: %f",ar);
     return 0;
     }
